backend/assignment1/requirements.txt
Django==4.1.7
gunicorn==20.1.0
Faker==18.4.0
django-cors-headers==3.14.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
uritemplate==4.1.1
urllib3==1.26.15
daphne==4.0.0
channels==4.0.0
channels-redis==4.1.0
pyyaml==6.0